+++ src/main/java/com/f5/onepageresumebe/domain/service/ProjectService.java
@@ -6,6 +6,7 @@ import com.f5.onepageresumebe.domain.entity.*;
 import com.f5.onepageresumebe.domain.repository.*;
 import com.f5.onepageresumebe.domain.repository.querydsl.ProjectQueryRepository;
 import com.f5.onepageresumebe.domain.repository.querydsl.UserQueryRepository;
+import com.f5.onepageresumebe.exception.customException.CustomException;
 import com.f5.onepageresumebe.exception.customException.CustomAuthenticationException;
 import com.f5.onepageresumebe.security.SecurityUtil;
 import com.f5.onepageresumebe.util.GitUtil;
@@ -30,6 +31,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import static com.f5.onepageresumebe.exception.ErrorCode.INVALID_INPUT_ERROR;
 
 @RequiredArgsConstructor //롬북을 통해서 간단하게 생성자 주입 방식의 어노테이션으로 fjnal이 붙거나 @notNull이 붙은 생성자들을 자동 생성해준다.
 @Service
@@ -61,15 +64,28 @@ public class ProjectService {
 
         projectRepository.save(project);
 
+        List<String> stacks = requestDto.getStack();
+
+        if(stacks.size()<3){
+            throw new CustomException("프로젝트 기술 스택을 3가지 이상 선택해 주세요.", INVALID_INPUT_ERROR);
+        }else if(multipartFiles.isEmpty()){
+            throw new CustomException("프로젝트 이미지를 1개 이상 업로드 해주세요.", INVALID_INPUT_ERROR);
+        }
+
         //스택 넣기
-        insertStacksInProject(project,requestDto.getStack());
+        insertStacksInProject(project, stacks);
 
         //이미지 넣기
         addImages(project,multipartFiles);
 
+        Integer projectId = project.getId();
+
         return ProjectResponseDto.builder()
-                .id(project.getId())
+                .id(projectId)
                 .title(project.getTitle())
+                .imageUrl(projectImgRepository.findFirstByProjectId(projectId).get()
+                        .getImageUrl())
+                .stack(stacks)
                 .build();
     }
 
@@ -110,13 +126,19 @@ public class ProjectService {
             throw new IllegalArgumentException("내가 작성한 프로젝트만 수정할 수 있습니다");
         }
 
+        List<String> stacks = requestDto.getStack();
+
+        if (stacks.size()<3){
+            throw new CustomException("프로젝트 스택을 3개 이상 선택해 주세요.",INVALID_INPUT_ERROR);
+        }
+
         project.updateIntro(requestDto);
 
         //기존에 있던 모든 연결된 스택 제거
         projectStackRepository.deleteAllByProjectId(projectId);
 
         //새로 들어온 스택 모두 프로젝트와 연결
-        insertStacksInProject(project, requestDto.getStack());
+        insertStacksInProject(project, stacks);
     }
 
     public List<ProjectResponseDto> getShortInfos(){
@@ -125,29 +147,7 @@ public class ProjectService {
 
         List<Project> projects = projectQueryRepository.findAllByUserEmail(email);
 
-        List<ProjectResponseDto> responseDtos = new ArrayList<>();
-
-        projects.forEach(project -> {
-            Integer projectId = project.getId();
-            String imageUrl = null;
-            ProjectImg projectImg = projectImgRepository.findFirstByProjectId(projectId).orElse(null);
-            if(projectImg!=null){
-                imageUrl = projectImg.getImageUrl();
-            }
-
-            ProjectResponseDto responseDto = ProjectResponseDto.builder()
-                    .id(projectId)
-                    .imageUrl(imageUrl)
-                    .title(project.getTitle())
-                    .bookmarkCount(project.getBookmarkCount())
-                    .stack(projectStackRepository.findStackNamesByProjectId(projectId))
-                    .content(project.getIntroduce())
-                    .build();
-
-            responseDtos.add(responseDto);
-        });
-
-        return responseDtos;
+        return ProjectUtil.projectToResponseDtos(projects, projectImgRepository, projectStackRepository);
     }
 
     public List<ProjectResponseDto> getAllByStacks(StackDto requestDto){
@@ -326,31 +326,3 @@ public class ProjectService {
         return projectDetailResponseDto;
      }
 }
-//
-//    String userEmail = SecurityUtil.getCurrentLoginUserId();
-//    User user = userQueryRepository.findByEmail(userEmail)
-//            .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 email 입니다."));
-//
-//    //프로젝트 주인이 조회하는지 체크
-//    Project project = projectQueryRepository.findByUserEmailAndProjectId(userEmail, projectId).orElse(null);
-//
-//    boolean isMyProject = true;
-//    boolean isBookmarking = false;
-//
-////본 주인이 아니라면, isMyProject : false, 프로젝트 가져오기
-//        if(project == null) {
-//                isMyProject = false;
-//                project = projectRepository.getById(projectId);
-//
-//                Optional<ProjectBookmark> optionalProjectBookmark = projectBookmarkRepository.findFirstByUserIdAndProjectId(user.getId(), projectId);
-//        isBookmarking = optionalProjectBookmark.isPresent();
-//        }
-//
-//
-//        ProjectDetailResponseDto projectDetailResponseDto = ProjectUtil.projectToDeatilResponseDto(project,
-//        projectImgRepository,
-//        projectStackRepository);
-//
-//        projectDetailResponseDto.checkBookmark(isMyProject, isBookmarking);
-//
-//        return projectDetailResponseDto;
